AdressParser.java: Added a getGroupNames() method which returns the String[] containing the group names for the indexes in the result of parseAdress(). This is to be used for debugging purposes e.g. when you are sysoing a result from parseAdress and want to match the results with the group names.
Main.java:
Changed to now also print the group names for each index in the ArrayList which parseAdress returns.

AdressParserTest.java:
Changed so it now tests the return of an ArrayList<ArrayList<String>>  instead of the String[] which parseAdress used to return.
Test cases remain the same.... TODO review test cases